# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Ordear Rest API Documentation",
    "description": "Backend Server Documentation Rest APIs",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:5555",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
"/avis/add/avis": {
"post": {
  "tags": [
    "Review"
  ],
  "summary": "Add a new review",
  "description": "Adds a new review for a restaurant and order.",
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "parameters": [
    {
      "name": "Authorization",
      "in": "cookie",
      "required": true,
      "description": "Auth Token",
      "type": "string"
    },
    {
      "name": "body",
      "in": "body",
      "required": true,
      "schema": {
        "type": "object",
        "properties": {
          "comment": {
            "type": "string",
            "description": "The review comment."
          },
          "note": {
            "type": "number",
            "description": "The review rating (e.g., 1 to 5)."
          },
          "restaurantFK": {
            "type": "string",
            "description": "The ID of the restaurant."
          },
          "orderFK": {
            "type": "string",
            "description": "The ID of the order."
          }
        },
        "required": [
          "comment",
          "note",
          "restaurantFK",
          "orderFK"
        ]
      }
    }
  ],
  "responses": {
    "201": {
      "description": "Review added successfully"
    }
  }
}
},
"/avis/getAllAvis": {
"parameters": [
  {
    "name": "Authorization",
    "in": "cookie",
    "required": true,
    "description": "Auth Token",
    "type": "string"
  }
],
"get": {
  "tags": [
    "Review"
  ],
  "summary": "Fetch all Review",
  "responses": {
    "200": {
      "description": "OK"
    }
  }
}
},
"/avis/getById/{id}": {
"get": {
  "tags": [
    "Review"
  ],
  "summary": "Get Review by Id",
  "parameters": [
    {
      "name": "id",
      "in": "path",
      "required": true,
      "description": "ID of the Review",
      "type": "string"
    }
  ],
  "responses": {
    "200": {
      "description": "OK"
    }
  }
}
},
"/avis/getByUser": {
"get": {
  "tags": [
    "Review"
  ],
  "summary": "Get Review by User",
  "parameters": [
    {
      "name": "Authorization",
      "in": "header",
      "required": true,
      "description": "Auth Token",
      "type": "string"
    }
  ],
  "responses": {
    "200": {
      "description": "OK"
    }
  }
}
},
"/avis/responseAvis/{id}": {
"parameters": [
  {
    "name": "id",
    "in": "path",
    "required": true,
    "description": "ID of Review",
    "type": "string"
  }
],
"put": {
  "summary": "Response Review ",
  "tags": [
    "Review"
  ],
  "parameters": [
    {
      "name": "avis",
      "in": "path",
      "required": true,
      "description": "ID of avis",
      "type": "string"
    },
    {
      "name": "update",
      "in": "body",
      "description": "Review response"
    }
  ],
  "responses": {
    "200": {
      "description": "Review response"
    }
  }
}
},
"/avis/getComments/{idAvis}": {
"parameters": [
  {
    "name": "idAvis",
    "in": "path",
    "required": true,
    "description": "ID of the Review",
    "type": "string"
  }
],
"get": {
  "tags": [
    "Review"
  ],
  "summary": "Get Comment by Id",
  "responses": {
    "200": {
      "description": "OK"
    }
  }
}
},
"/avis/addComment/{idAvis}": {
"post": {
  "tags": [
    "Review"
  ],
  "summary": "Add new Comment",
  "parameters": [
    {
      "name": "idAvis",
      "in": "path",
      "required": true,
      "description": "The ID of the review to which the comment is being added",
      "type": "string"
    },
    {
      "name": "body",
      "in": "body",
      "required": true,
      "schema": {
        "type": "object",
        "properties": {
          "comment": {
            "type": "string",
            "description": "The comment to be added"
          },
          "restaurantFK": {
            "type": "string",
            "description": "The ID of the restaurant related to the review"
          }
        }
      }
    }
  ],
  "produces": [
    "application/json"
  ],
  "responses": {
    "200": {
      "description": "New comment is added"
    }
  }
}
},
"/cart/addProdductToCart": {
"post": {
  "tags": [
    "Cart"
  ],
  "summary": "Add item to user's cart",
  "description": "Adds a product to the user's cart for later order.",
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "parameters": [
    {
      "name": "Authorization",
      "in": "cookie",
      "required": true,
      "description": "Auth Token",
      "type": "string"
    },
    {
      "name": "body",
      "in": "body",
      "required": true,
      "schema": {
        "type": "object",
        "properties": {
          "productFK": {
            "type": "string",
            "description": "The ID of the product to add to the cart."
          },
          "ingredientFK": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "An array of ingredient IDs associated with the product."
          },
          "itemsFK": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "An array of item IDs associated with the product."
          },
          "tableNb": {
            "type": "string",
            "description": "The table number associated with the cart item."
          },
          "restaurantFK": {
            "type": "string",
            "description": "The ID of the restaurant."
          }
        },
        "required": [
          "productFK",
          "ingredientFK",
          "itemsFK",
          "tableNb",
          "restaurantFK"
        ]
      }
    }
  ],
  "responses": {
    "200": {
      "description": "Item added to cart"
    }
  }
}
},
"/cart/get/cartTrash/by/user": {
"get": {
  "tags": [
    "Cart"
  ],
  "summary": "Get Cart Trash by User",
  "parameters": [
    {
      "name": "Authorization",
      "in": "header",
      "required": true,
      "description": "Auth Token",
      "type": "string"
    }
  ],
  "responses": {
    "200": {
      "description": "OK"
    }
  }
}
},
"/cart/get/cartOrder/by/user": {
"get": {
  "tags": [
    "Cart"
  ],
  "summary": "Get Cart Order by User",
  "parameters": [
    {
      "name": "Authorization",
      "in": "header",
      "required": true,
      "description": "Auth Token",
      "type": "string"
    }
  ],
  "responses": {
    "200": {
      "description": "OK"
    }
  }
}
},
"/cart/remove/{id}": {
"delete": {
  "tags": [
    "Cart"
  ],
  "summary": "Delete product from cart",
  "parameters": [
    {
      "name": "id",
      "in": "path",
      "required": true,
      "description": "ID od the product",
      "type": "string"
    }
  ],
  "responses": {
    "200": {
      "description": "OK"
    }
  }
}
},
"/cart/clear": {
"delete": {
  "tags": [
    "Cart"
  ],
  "parameters": [
    {
      "name": "Authorization",
      "in": "header",
      "required": true,
      "description": "Auth Token",
      "type": "string"
    }
  ],
  "summary": "Delete cart trash",
  "responses": {
    "200": {
      "description": "OK"
    }
  }
}
},
"/cart/increase/quantity/Cart/Trash/{idCartTrash}/{productFK}": {
  "put": {
    "tags": [
      "Cart"
    ],
    "summary": "Increase product quantity in cart trash",
    "parameters": [
      {
        "name": "productFK",
        "in": "path",
        "required": true,
        "type": "string"
      },
      {
        "name": "idCartTrash",
        "in": "path",
        "required": true,
        "type": "string"
      },
      {
        "name": "Authorization",
        "in": "header",
        "required": true,
        "description": "Auth Token",
        "type": "string"
      }
    ],
    "responses": {
      "200": {
        "description": "Product quantity is increased"
      },
      "404": {
        "description": "Not Product Found"
      },
      "500": {
        "description": "Internal Server Error"
      }
    }
  }
},
"/cart/increase/quantity/Cart/Order/{productFK}": {
  "put": {
    "tags": [
      "Cart"
    ],
    "summary": "Increase product quantity in cart Order",
    "parameters": [
      {
        "name": "productFK",
        "in": "path",
        "required": true,
        "type": "string"
      },
      {
        "name": "Authorization",
        "in": "cookie",
        "required": true,
        "description": "Auth Token",
        "type": "string"
      }
    ],
    "responses": {
      "200": {
        "description": "Product quantity is increased"
      },
      "404": {
        "description": "Not Product Found"
      },
      "500": {
        "description": "Internal Server Error"
      }
    }
  }
},
"/cart/decrease/quantity/Cart/Trash/{idCartTrash}/{productFK}": {
  "put": {
    "tags": [
      "Cart"
    ],
    "summary": "Decrease product quantity in cart trash",
    "parameters": [
      {
        "name": "productFK",
        "in": "path",
        "required": true,
        "type": "string"
      },
      {
        "name": "idCartTrash",
        "in": "path",
        "required": true,
        "type": "string"
      },
      {
        "name": "Authorization",
        "in": "header",
        "required": true,
        "description": "Auth Token",
        "type": "string"
      }
    ],
    "responses": {
      "200": {
        "description": "Product quantity is increased"
      },
      "404": {
        "description": "Not Product Found"
      },
      "500": {
        "description": "Internal Server Error"
      }
    }
  }
},
"/cart/decrease/quantity/Cart/Order/{productFK}": {
  "put": {
    "tags": [
      "Cart"
    ],
    "summary": "Decrease product quantity in cart Order",
    "parameters": [
      {
        "name": "productFK",
        "in": "path",
        "required": true,
        "type": "string"
      },
      {
        "name": "Authorization",
        "in": "header",
        "required": true,
        "description": "Auth Token",
        "type": "string"
      }
    ],
    "responses": {
      "200": {
        "description": "Product quantity is increased"
      },
      "404": {
        "description": "Not Product Found"
      },
      "500": {
        "description": "Internal Server Error"
      }
    }
  }
},
...
  },
...
}